Bug fixes and enhancements
o Fix Issue 85 - camera worker thread hang

 - worker threads should be doing CoInitializeEx(COINIT_MULTITRHEADED)
   since they do not have a windows message loop

 - it is not safe for worker threads to directly reference camera COM
   interface pointers instantiated in the main thread. Use the
   GlobalInterfaceTable to marshall the pointers just like we do with
   the ASCOM Mount interface pointers.

 - make sure that when worker threads call Yield they do not try to
   pump windows messages. Eventually we should remove all the Yield
   calls from code that runs in worker threads, but for now the Yield
   calls are still there but have no effect in the worker threads.

 - this change also Updates Issue 87: the ASCOM camera simulator no
   longer crashes PHD2 (it still does not work yet though)

o remove a bunch of casting by having
  pFrame->RequestedExposureDuration() return int

o AO Secondary mount bumping enhancements

    - always start a new bump when the AO is outside the safe range,
      rather than waiting for the current bump to complete

    - dynamically adjust the minumim bump increment so that bumping
      becomes more aggressive as long as the AO position is remaining
      outside the safe range

    - keep track of the recent average position of the AO and base the
      bump on the average position, not the current position

    - on the AO graph, display the average position, the current bump
      vector and the remaining bump vector

o fix an issue where an error during camera capture would casue PHD2
  to get stuck with "waiting for devices to stop"

+++ myframe.h
@@ -119,7 +119,7 @@ private:
     long m_instanceNumber;
 
     wxAuiManager m_mgr;
-    bool m_continueCapturing; // should anouther image be captuered?
+    bool m_continueCapturing; // should another image be captured?
 
 public:
     MyFrame(const wxString& title, int instanceNumber, wxLocale *locale);
@@ -201,7 +201,7 @@ public:
 
     bool StartServer(bool state);
     bool FlipRACal(wxCommandEvent& evt);
-    double RequestedExposureDuration();
+    int RequestedExposureDuration();
     bool Voyager_Connect();
     int GetFocalLength(void);
     int GetLanguage(void);
@@ -212,7 +212,7 @@ public:
     struct EXPOSE_REQUEST
     {
         usImage          *pImage;
-        double           exposureDuration;
+        int              exposureDuration;
         wxRect           subframe;
         bool             bError;
         wxSemaphore      *pSemaphore;
@@ -231,7 +231,7 @@ public:
     };
     void OnRequestMountMove(wxCommandEvent &evt);
 
-    void ScheduleExposure(double exposureDuration, wxRect subframe);
+    void ScheduleExposure(int exposureDuration, wxRect subframe);
     void SchedulePrimaryMove(Mount *pMount, const PHD_Point& vectorEndpoint, bool normalMove=true);
 
     void ScheduleSecondaryMove(Mount *pMount, const PHD_Point& vectorEndpoint, bool normalMove=true);
@@ -267,7 +267,7 @@ private:
     wxMessageQueue<STATUSBAR_QUEUE_ENTRY> m_statusbarQueue;
     wxTimer m_statusbarTimer;
 
-    double m_exposureDuration;
+    int m_exposureDuration;
 
     void OnSetStatusText(wxThreadEvent& event);
     void OnStatusbarTimerEvent(wxTimerEvent& evt);
@@ -281,7 +281,7 @@ private:
     void SetupHelpFile(void);
     void SetComboBoxWidth(wxComboBox *control, unsigned int extra);
 
-    double ExposureDurationFromSelection(const wxString& selection);
+    int ExposureDurationFromSelection(const wxString& selection);
 
     // and of course, an event table
     DECLARE_EVENT_TABLE()
